CodigoInterfazGrafica/requirements.txt
accelerate==0.21.0
aenum==3.1.16
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
blendmodes==2022
certifi==2025.4.26
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
clip==1.0
colorama==0.4.6
contourpy==1.3.2
control==0.10.1
cycler==0.12.1
deprecation==2.1.0
diskcache==5.6.3
einops==0.4.1
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.94.0
ffmpy==0.5.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.32
gradio==3.41.2
gradio_client==0.5.0
h11==0.12.0
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.6
jsonmerge==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kornia==0.6.7
lark==1.1.2
lazy_loader==0.4
lightning-utilities==0.14.3
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
narwhals==1.36.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.2
omegaconf==2.2.3
open-clip-torch==2.20.0
opencv-python==4.11.0.86
orjson==3.10.16
packaging==25.0
pandas==2.2.3
piexif==1.1.3
Pillow==9.5.0
pillow-avif-plugin==1.4.3
propcache==0.3.1
protobuf==3.20.0
psutil==5.9.5
pydantic==1.10.22
pydub==0.25.1
pyparsing==3.2.3
pyqtgraph==0.13.7
pyserial==3.5
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==1.9.4
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resize-right==0.0.2
rpds-py==0.24.0
safetensors==0.4.2
scikit-image==0.21.0
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
shiboken6==6.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
spandrel==0.3.4
spandrel_extra_arches==0.1.1
starlette==0.26.1
sympy==1.13.3
tifffile==2025.3.30
timm==1.0.15
tokenizers==0.13.3
tomesd==0.1.3
torch==2.1.2+cu121
torchdiffeq==0.2.3
torchmetrics==1.7.1
torchsde==0.2.6
torchvision==0.16.2+cu121
tqdm==4.67.1
trampoline==0.1.2
transformers==4.30.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==11.0.3
yarl==1.20.0
